+++ src/server.ts
@@ -9,6 +9,7 @@
  * rights grant found at http://polymer.github.io/PATENTS.txt
  */
 
+import {parse as babelParse} from '@babel/parser';
 import * as http from 'http';
 import * as net from 'net';
 import * as path from 'path';
@@ -18,13 +19,16 @@ import mount = require('koa-mount');
 import send = require('koa-send');
 import serve = require('koa-static');
 import bodyParser = require('koa-bodyparser');
+import {nodeResolve} from 'koa-node-resolve';
 
 import {BenchmarkResponse, Deferred} from './types';
 
 export interface ServerOpts {
   host: string;
   ports: number[];
+  root: string;
   mountPoints: MountPoint[];
+  resolveBareModules: boolean;
 }
 
 export interface MountPoint {
@@ -80,6 +84,22 @@ export class Server {
     app.use(mount('/submitResults', this.submitResults.bind(this)));
     app.use(this.instrumentRequests.bind(this));
     app.use(this.serveBenchLib.bind(this));
+
+    if (opts.resolveBareModules === true) {
+      app.use(nodeResolve({
+        root: opts.root,
+        // Only log errors.
+        logger: {...console, debug: undefined, info: undefined},
+        // Enable latest JS syntax.
+        jsParser: (js) => babelParse(js, {
+          sourceType: 'unambiguous',
+          plugins: [
+            'dynamicImport',
+            'importMeta',
+          ],
+        }),
+      }));
+    }
     for (const {diskPath, urlPath} of opts.mountPoints) {
       app.use(mount(urlPath, serve(diskPath, {index: 'index.html'})));
     }